+++ cpp/pep/storagefacility/PageStore.cpp
@@ -24,6 +24,8 @@
 #include <prometheus/gauge.h>
 #include <prometheus/registry.h>
 
+using namespace std::ranges;
+
 namespace pep
 {
 
@@ -31,11 +33,22 @@ namespace {
 
   const std::string LogTag("PageStore");
 
+  //region S3PageStore
   class S3PageStore
     : public PageStore,
       public std::enable_shared_from_this<S3PageStore>
   {
   public:
+    struct HostParameters {
+      s3::Client::Parameters clientParams;
+      unsigned connections;
+    };
+
+    struct Bucket {
+      std::string name;
+      std::string hostId;
+      [[nodiscard]] bool operator==(const Bucket&) const = default;
+    };
 
     messaging::MessageSequence
       get(const std::string& path) override;
@@ -51,29 +64,40 @@ namespace {
 
     // pubic constructor for the sake of std::make_shared
     S3PageStore(
-      const s3::Client::Parameters& s3params,
-        unsigned int conn_count,
-        const std::string& writeBucket_,
-        const std::vector<std::string>& buckets_,
+        const std::unordered_map<std::string, HostParameters>& hostsParams,
+        std::vector<Bucket> readBuckets,
+        Bucket writeBucket,
         std::shared_ptr<prometheus::Registry> metrics_registry);
 
     ~S3PageStore() override;
 
   private:
-    std::vector<std::shared_ptr<s3::Client>> clients_;
+    static Bucket ParseBucket(const Configuration& config) {
+      return {
+        .name = config.get<std::string>("Name"),
+        .hostId = config.get<std::string>("HostId"),
+      };
+    }
 
-    // keeps track of the number of open requests per connection
-    std::shared_ptr<std::vector<unsigned int>> openRequestsCounts_;
+    struct Connection {
+      std::shared_ptr<s3::Client> client;
+      // keeps track of the number of open requests per connection
+      unsigned int openRequestsCounts{};
+    };
+    struct Host {
+      std::vector<Connection> connections;
+      Connection& quietestConnection() {
+        return *min_element(connections, {}, &Connection::openRequestsCounts);
+      }
+    };
+    std::unordered_map<std::string, Host> hosts_;
 
-    std::string writeBucket_;
-    std::vector<std::string> buckets_;
-
-    // gets the index of (one of) the quietest connections
-    size_t getQuietestConn();
+    std::vector<Bucket> readBuckets_;
+    Bucket writeBucket_;
 
     // gets page from specified bucket
     messaging::MessageSequence get(const std::string& path,
-        const std::string bucket);
+        const Bucket& bucket);
 
 
     struct Metrics {
@@ -108,100 +132,106 @@ namespace {
       std::shared_ptr<prometheus::Registry> metrics_registry,
       const Configuration& config)
   {
-    s3::Client::Parameters s3params = {
-      config.get<EndPoint>("EndPoint"),
-      config.get<s3::Credentials>("Credentials"),
-      io_context,
-      config.get<std::optional<std::filesystem::path>>("CaCertificateFile"),
-      config.get<std::optional<bool>>("UseHttps")
-    };
-
-    unsigned int conn_count = config.get<unsigned int>("Connections", 5);
-    std::string writeBucket = config.get<std::string>("WriteToBucket");
-
-    std::vector<std::string> buckets;
-
-    for (const std::string& bucket
-        : config.get<std::vector<std::string>>("ReadFromBuckets")) {
-      buckets.push_back(bucket);
-    }
-
-    if (buckets.begin() == buckets.end())
-      throw std::runtime_error("S3PageStore configuration error: "
-          "no buckets_ to read from!");
-
-    if (std::find(buckets.begin(), buckets.end(), writeBucket)
-          == buckets.end()) {
-      throw std::runtime_error("S3PageStore configuration error: "
-          "writing to a bucket we're not reading from!");
-    }
+    auto hostsParams = RangeToCollection<std::unordered_map<std::string, HostParameters>>(
+      config.get_children_map("Hosts")
+      | views::transform([&io_context](const auto& entry) {
+        const Configuration& hostConfig = entry.second;
+        return std::pair{
+          entry.first,
+          HostParameters{
+            .clientParams = {
+              .endpoint = hostConfig.get<EndPoint>("EndPoint"),
+              .credentials = hostConfig.get<s3::Credentials>("Credentials"),
+              .ioContext = io_context,
+              .caCertPath = hostConfig.get<std::optional<std::filesystem::path>>("CaCertificateFile"),
+              .useHttps = hostConfig.get<std::optional<bool>>("UseHttps")
+            },
+            .connections = hostConfig.get<unsigned int>("Connections", 5),
+          },
+        };
+      }));
 
     return std::make_shared<S3PageStore>(
-        s3params, conn_count, writeBucket, buckets, metrics_registry);
+      hostsParams,
+      RangeToVector(config.get_children_vector("ReadFromBuckets")
+        | views::transform(ParseBucket)),
+      ParseBucket(config.get_child("WriteToBucket")),
+      metrics_registry);
   }
 
 
 
   S3PageStore::S3PageStore(
-      const s3::Client::Parameters& s3params,
-      unsigned int conn_count,
-      const std::string& writeBucket_,
-      const std::vector<std::string>& buckets_,
+      const std::unordered_map<std::string, HostParameters>& hostsParams,
+      std::vector<Bucket> readBuckets,
+      Bucket writeBucket,
       std::shared_ptr<prometheus::Registry> metrics_registry)
 
-    : PageStore(),
-      clients_(),
-      openRequestsCounts_(std::make_shared<std::vector<unsigned int>>()),
-      writeBucket_(writeBucket_),
-      buckets_(buckets_),
+    : readBuckets_(std::move(readBuckets)),
+      writeBucket_(std::move(writeBucket)),
       metrics_(metrics_registry ? std::make_optional<Metrics>(metrics_registry)
                                : std::nullopt)
   {
-    for (auto i = 0U; i < conn_count; i++) {
-      auto client = s3::Client::Create(s3params);
-      client->start();
-      this->clients_.push_back(client);
-      this->openRequestsCounts_->push_back(0);
+    if (readBuckets_.empty()) {
+      throw std::runtime_error("S3PageStore configuration error: "
+          "no buckets to read from!");
+    }
+    if (find(readBuckets_, writeBucket_) == readBuckets_.end()) {
+      throw std::runtime_error("S3PageStore configuration error: "
+          "writing to a bucket we're not reading from!");
+    }
+
+    const auto initializeHost = [&](const std::string& id) {
+      const auto [hostIt, emplaced] = hosts_.try_emplace(id);
+      if (emplaced) {
+        const auto hostParamsIt = hostsParams.find(id);
+        if (hostParamsIt == hostsParams.end()) {
+          throw std::runtime_error("S3PageStore configuration error: "
+              "referenced host \"" + id + "\" not found in configuration");
+        }
+        const HostParameters& hostParams = hostParamsIt->second;
+        if (hostParams.connections == 0) {
+          throw std::runtime_error("S3PageStore configuration error: "
+              "number of connections for a host must be nonzero");
+        }
+        hostIt->second = Host{
+          .connections = RangeToVector(
+          views::iota(0u, hostParams.connections)
+          | views::transform([&](unsigned int) {
+            auto client = s3::Client::Create(hostParams.clientParams);
+            client->start();
+            return Connection{std::move(client)};
+          })),
+        };
+      }
+    };
+
+    for (const Bucket& bucket : readBuckets_) {
+      initializeHost(bucket.hostId);
+    }
+    initializeHost(writeBucket_.hostId);
+
+    for (const std::string& hostId : views::keys(hostsParams)) {
+      if (!hosts_.contains(hostId)) {
+        PEP_LOG(LogTag, Severity::Warning) << "Host defined but not referenced: " << hostId;
+      }
     }
   }
 
   S3PageStore::~S3PageStore() {
-    for (auto client : clients_) {
-      client->shutdown();
-    }
-#if PEP_BUILD_HAS_DEBUG_FLAVOR()
-    for (unsigned int count : *(this->openRequestsCounts_))
-      assert(count == 0);
-    // The "count" variable is only used in an assertion, making it
-    // unused in non-debug builds.
-    //
-    // Why not a PEP_LOG(LogTag, Severity::Error) here instead of an assert?
-    //
-    // Either there's a bug in the open requests counting code---which we don't
-    // want to be buried in the logs---or some request is actually still active,
-    // which will cause an inexplicable segfault when it'll try to decrement
-    // the deleted openRequestsCounts_[idx] upon completion.
-#endif
-  }
+    for (const auto& host : views::values(hosts_)) {
+      for (const auto& connection : host.connections) {
+        connection.client->shutdown();
 
-
-  size_t S3PageStore::getQuietestConn() {
-    size_t candidate = 0;
-
-    unsigned int candidate_count = this->openRequestsCounts_->at(candidate);
-
-    for (size_t contender=1;
-        contender < this->openRequestsCounts_->size(); contender++) {
-
-      unsigned int contender_count = this->openRequestsCounts_->at(contender);
-
-      if (contender_count < candidate_count) {
-        candidate = contender;
-        candidate_count = contender_count;
+        // Why not a PEP_LOG(LogTag, Severity::Error) here instead of an assert?
+        //
+        // Either there's a bug in the open requests counting code---which we don't
+        // want to be buried in the logs---or some request is actually still active,
+        // which will cause an inexplicable segfault when it'll try to decrement
+        // the deleted openRequestsCounts upon completion.
+        assert(connection.openRequestsCounts == 0);
       }
     }
-
-    return candidate;
   }
 
 
@@ -238,7 +268,7 @@ namespace {
     messaging::MessageSequence result
       = rxcpp::observable<>::empty<std::shared_ptr<std::string>>();
 
-    for (const std::string& bucket : this->buckets_) {
+    for (const Bucket& bucket : this->readBuckets_) {
 
       result = result.switch_if_empty(RxLazy<std::shared_ptr<std::string>>(
 
@@ -256,9 +286,8 @@ namespace {
 
 
   messaging::MessageSequence S3PageStore::get(
-      const std::string& _path, std::string bucket)
+      const std::string& path, const Bucket& bucket)
   {
-    std::string path = _path; // don't leave the reference dangling
     auto self = this->shared_from_this();
 
     if (metrics_)
@@ -277,30 +306,32 @@ namespace {
         self->metrics_->pending_requests.Decrement();
     });
 
+    Host& host = hosts_.at(bucket.hostId);
+
     // The "subscribe" on the returned observable may be called much later,
     // so we do not immediately pick a connection.
     return RxLazy<std::shared_ptr<std::string>>(
-    [self,path,bucket,post_pending=std::move(post_pending)]()
+    [self, path, &host, bucket = bucket.name, post_pending=std::move(post_pending)]()
       -> messaging::MessageSequence {
 
-      size_t conn_idx = self->getQuietestConn();
+      Connection& connection = host.quietestConnection();
 
       post_pending->trigger();
       // NB. We can't use post_pending.reset() since post_pending is const.
 
-      (self->openRequestsCounts_->at(conn_idx))++;
+      ++connection.openRequestsCounts;
 
       if (self->metrics_) {
         self->metrics_->active_requests.Increment();
       }
 
-      auto post_active = DeferShared([self, conn_idx]{
-        (self->openRequestsCounts_->at(conn_idx))--;
+      auto post_active = DeferShared([self, &connection]{
+        --connection.openRequestsCounts;
         if (self->metrics_)
           self->metrics_->active_requests.Decrement();
       });
 
-      return self->clients_.at(conn_idx)->getObject(path, bucket)
+      return connection.client->getObject(path, bucket)
         .op(RxButFirst(
 
           // RxButFirst makes sure the function below is called after
@@ -315,10 +346,9 @@ namespace {
 
 
   rxcpp::observable<std::string> S3PageStore::put(
-      const std::string& _path,
+      const std::string& path,
       std::vector<std::shared_ptr<std::string>> page_parts)
   {
-    std::string path = _path; // don't leave the reference dangling
     size_t pages_size = 0;
 
     for (const auto& page_part : page_parts) {
@@ -341,29 +371,31 @@ namespace {
       }
     });
 
+    Host& host = hosts_.at(writeBucket_.hostId);
+
     // The "subscribe" on the returned observable may be called much later,
     // so we do not immediately pick a connection.
     return RxLazy<std::string>(
-    [self,path,page_parts=std::move(page_parts),post_pending=std::move(post_pending)]()
+    [self,path,&host,page_parts=std::move(page_parts),post_pending=std::move(post_pending)]()
       -> rxcpp::observable<std::string> {
 
-      size_t conn_idx = self->getQuietestConn();
+      Connection& connection = host.quietestConnection();
 
       post_pending->trigger();
 
-      (self->openRequestsCounts_->at(conn_idx))++;
+      ++connection.openRequestsCounts;
       if (self->metrics_) {
         self->metrics_->active_requests.Increment();
       }
 
-      auto post_active = DeferShared([self,conn_idx](){
-        (self->openRequestsCounts_->at(conn_idx))--;
+      auto post_active = DeferShared([self,&connection](){
+        --connection.openRequestsCounts;
         if (self->metrics_)
           self->metrics_->active_requests.Decrement();
       });
 
-      return self->clients_.at(conn_idx)->putObject(path,
-        self->writeBucket_, page_parts)
+      return connection.client->putObject(path,
+        self->writeBucket_.name, page_parts)
         .op(RxButFirst(
 
           // RxButFirst makes sure the function below is called after
@@ -375,7 +407,9 @@ namespace {
 
     });
   }
+  //endregion S3PageStore
 
+  //region LocalPageStore
   // stores data directly on disk
   class LocalPageStore
     : public PageStore,
@@ -475,8 +509,10 @@ namespace {
         s.on_completed();
       });
   }
+  //endregion LocalPageStore
 
 
+  //region DualPageStore
   // Run both a LocalPageStore and an S3PageStore - to see if they agree.
   class DualPageStore
     : public PageStore,
@@ -594,6 +630,7 @@ namespace {
         }
       }).as_dynamic();
   }
+  //endregion DualPageStore
 
 }